ipa: Use pseudo-random number for SLS in IPA->M3UA direction
[back-port of Ia4e66d660b6057338f66a47fffc8a0d32759f733]

In Change-Id Ice7bab997b84cfed00c7d6d780c70f4e9fac6002 we introduced
code that would make the LSB of the file descriptor be used as SLS
when passing packets from IPA in M3UA direction.

This did however not achieve sufficient entropy in real-world use cases.

In this change, we change over to allocating a pseudo-random SLS to each
IPA connection at the time it is established;  We then assign that SLS
to each packet received on that IPA connection.

Change-Id: I97f43bd3bdd89a6bf21a84284a060305fcbb86e7
Related: SYS#6543
Closes: SYS#6802