xua_asp_fsm: Fix 'Event AS_ASSIGNED not permitted'
The event case handling (ignore it) was added in a commit some time ago,
but it forgot to add it to the allstate_event_mask in the xua_fsm (it
was added only in the ipa_fsm, which is the one using it).

This bug should in theory be harmless since the only code dispatching
the signal is not checking the return code of osmo_fsm_inst_dispatch().

Related: SYS#5914
Fixes: 57444690212b1576eaf0a7156448ba2d27c59050
Change-Id: Iaead46bbc40b923763bc3dbe4d24d8952822de4a