remove old static libsccp.a, libmtp.a, libxua.a
Those static libraries were not used or maintained a long time ago.

We have a much more complete implementation of SCCP and xUA in
libosmo-sigtran.  That has been the basis of osmo-stp, as well as in
osmo-sgsn.git, osmo-bsc.git, osmo-msc.git.

The only user of the old static libraries was osmo-bsc-sccplite from
openbsc.git, which is unsupported for many years.

Change-Id: I299a32a2f8f61477cf49c8126567fef2092d5407